Create hook to separate worker and resize ligic
