+++ src/types.ts
@@ -4,6 +4,11 @@ export type AnchorObject = {
   blob: Blob
 }
 
+export interface FileWithBlob {
+  blob: Blob
+  name: string
+}
+
 export enum ResizeState {
   TO_LOAD = 'TO_LOAD',
   LOADED = 'LOADED',
